+++ Source/ChocolateyGui/Services/IChocolateyPackageService.cs
@@ -7,6 +7,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using ChocolateyGui.Models;
 using ChocolateyGui.ViewModels.Items;
 using NuGet;
 
@@ -14,6 +15,14 @@ namespace ChocolateyGui.Services.PackageServices
 {
     public interface IChocolateyPackageService
     {
+        Task<PackageSearchResults> Search(string query);
+
+        Task<PackageSearchResults> Search(string query, PackageSearchOptions options);
+
+        Task<IPackageViewModel> GetLatest(string id, bool includePrerelease = false);
+
+        Task<IPackageViewModel> GetByVersionAndIdAsync(string id, SemanticVersion version, bool isPrerelease);
+
         Task<IEnumerable<IPackageViewModel>> GetInstalledPackages(bool force = false);
 
         Task InstallPackage(string id, SemanticVersion version = null, Uri source = null, bool force = false);
@@ -21,5 +30,9 @@ namespace ChocolateyGui.Services.PackageServices
         Task UninstallPackage(string id, SemanticVersion version, bool force = false);
 
         Task UpdatePackage(string id, Uri source = null);
+
+        Task PinPackage(string id, SemanticVersion version);
+
+        Task UnpinPackage(string id, SemanticVersion version);
     }
 }
