+++ Source/ChocolateyGui/packages.config
@@ -11,6 +11,7 @@
   <package id="log4net" version="2.0.3" targetFramework="net45" />
   <package id="MahApps.Metro" version="1.3.0" targetFramework="net45" />
   <package id="MahApps.Metro.IconPacks.Entypo" version="1.6.0" targetFramework="net45" />
+  <package id="MahApps.Metro.IconPacks.Modern" version="1.6.0" targetFramework="net45" />
   <package id="MahApps.Metro.Resources" version="0.6.1.0" targetFramework="net45" />
   <package id="Mindscape.Raygun4Net" version="5.4.0" targetFramework="net45" />
   <package id="Rx-Core" version="2.2.5" targetFramework="net45" />
